[APP-858] re-organize data layer in apps feature module.
