+++ feature_apps/src/main/java/cm/aptoide/pt/feature_apps/data/network/model/WidgetType.java
@@ -1,4 +1,4 @@
-package cm.aptoide.pt.feature_apps.data;
+package cm.aptoide.pt.feature_apps.data.network.model;
 
 public enum WidgetType {
   APPS_GROUP, ADS, APPCOINS_ADS, ESKILLS, ACTION_ITEM, NEWS_ITEM, NEW_APP, NEW_APP_VERSION, IN_GAME_EVENT, APP_COMING_SOON,
