+++ feature_apps/src/main/java/cm/aptoide/pt/feature_apps/data/network/service/WidgetsRemoteService.kt
@@ -1,5 +1,7 @@
-package cm.aptoide.pt.feature_apps.data
+package cm.aptoide.pt.feature_apps.data.network.service
 
+import cm.aptoide.pt.feature_apps.data.network.model.BaseV7DataListResponse
+import cm.aptoide.pt.feature_apps.data.network.model.WidgetsJSON
 import retrofit2.Response
 import retrofit2.http.GET
