+++ feature_apps/src/main/java/cm/aptoide/pt/feature_apps/data/network/service/WidgetsRetrofitService.kt
@@ -1,4 +1,4 @@
-package cm.aptoide.pt.feature_apps.data
+package cm.aptoide.pt.feature_apps.data.network.service
 
 internal interface WidgetsRetrofitService : WidgetsRemoteService {
